+++ stub/corfu.el
@@ -4,7 +4,9 @@
 
 (defvar corfu--candidates nil)
 (defvar corfu--index nil)
+(defvar corfu--initial-state nil)
 (defvar corfu--input nil)
+;; Removed https://github.com/minad/corfu/commit/63d1de2696adcb09a4ea01ba668635364e37a9c2
 (defvar corfu--state-vars nil)
 (defvar corfu-map nil)
 (defvar corfu-mode nil)