+++ vertico-prescient.el
@@ -8,7 +8,7 @@
 ;; Created: 23 Sep 2022
 ;; Package-Requires: ((emacs "27.1") (prescient "6.1.0") (vertico "0.28") (compat "29.1"))
 ;; SPDX-License-Identifier: MIT
-;; Version: 6.2.0
+;; Version: 6.3.0
 
 ;;; Commentary:
 
@@ -262,7 +262,7 @@ This mode will:
     (when (equal (lookup-key vertico-map (kbd "M-s"))
                  prescient-toggle-map)
       (define-key vertico-map (kbd "M-s")
-                  vertico-prescient--old-toggle-binding))
+        vertico-prescient--old-toggle-binding))
     (remove-hook 'prescient--toggle-refresh-functions
                  #'vertico-prescient--toggle-refresh)